Add shift-click support to AdvancedFilterTable
This PR also breaks AdvancedFilterTable up into smaller files
